Provide for a findSimilar that works across the collections on distributed machines.  This means that we need to do a two step process whereby we get the document vector for the document from the collection where it is stored and then use that document vector to run the findSimilars in each of the collections.  This required us to add a new interface LowLevelSearch that is used by PartitionCluster and Replicant, since the methods at the DataStore level are now different than those used at these levels (due to the two-steppiness).
Also added a query method to the ItemSearch interface so that we can run queries directly on the search index.  This is meant to be temporary, but is likely to be temporary like the buildings in Ottawa that were built for WWII and are still being used.

Added commands for querying and findSimilaring to the shell.