A new set of configurations for a "coarse" distribution (e.g., on a single machine) so that we can restart parts as necessary.
Changed all of the configurations to use the auraGroup global variable for the group to use to look for a registrar so that we can avoid registrar collisions while testing.

Added the auraPolicy global variable so that the policy file can be specified on the command line for the VM.

Removed confusing, redundant RecommenderManager.